requirements.txt
fastapi
uvicorn
jinja2
python-multipart
langchain
langgraph
langchain-openai
pydantic-settings
python-docx
